+++ MouseFinder.cpp
@@ -10,12 +10,12 @@ MouseFinder::MouseFinder()
 	}, this);
 }
 
-cv::Point2f MouseFinder::LocateOnScreen(HSVColorRangeMap &HSVRanges, cv::Mat &frameHSV, cv::Mat &frameBGR, OBJECT target)
+cv::Point2i MouseFinder::LocateOnScreen(HSVColorRangeMap &HSVRanges, cv::Mat &frameHSV, cv::Mat &frameBGR, OBJECT target)
 {
 	cv::Scalar colorCircle(133, 33, 55);
 	cv::circle(frameBGR, cv::Point2i(mouseLocation), 10, colorCircle, 3);
 	cv::imshow("MouseFinder", frameBGR);
-	return cv::Point2f(mouseLocation.x, mouseLocation.y);
+	return cv::Point2i(mouseLocation.x, mouseLocation.y);
 
 }
 MouseFinder::~MouseFinder()
